Used to address grievances or queries
A “Grievance/Complaint” is defined as any communication that expresses dissatisfaction about an action or lack of action, about the standard of service/deficiency of service of an institute and the complainant
asks for remedial action.Grievance Redress Mechanism is part and parcel of the machinery of any administration.  This project is such an outcome of our college grievance redressal system in 5 categories namely academics, sports, events, hostel and mess and other issues.